+++ lwk_bindings/tests/bindings/bip44discovery.py
@@ -16,15 +16,17 @@ w.wait_for_tx(txid, client)
 # Perform "BIP44 account discovery"
 
 # Get values shared by all (singlesig) descriptors
-# ss_desc_from_external_signer() wants the raw SLIP77 key, not wrapped in "slip77(...)"
-master_blinding_key = signer.slip77_master_blinding_key()[len("slip77("):-1]
+master_blinding_key = signer.slip77_master_blinding_key()  # "slip77(...)"
 fingerprint = signer.fingerprint()
 
 def derive_descriptor(account_type, account_num):
     account_path = DerivationPath.ss_path(network, account_type, account_num)
     # Get the xpub from the signer (which might be handled externally)
-    account_xpub = signer.keyorigin_xpub_from_path(account_path).split("]")[1]  # strip keyorigin
-    return WolletDescriptor.ss_desc_from_external_signer(network, account_type, account_num, master_blinding_key, fingerprint, account_xpub)
+    keyorigin_xpub = signer.keyorigin_xpub_from_path(account_path)
+    # Some signers just return the xpub, reconstruct the "keyorigin xpub" string
+    account_xpub = keyorigin_xpub.split("]")[1]  # strip keyorigin
+    dpk = DescriptorPublicKey(f"[{fingerprint}/{account_path}]{account_xpub}")
+    return WolletDescriptor.ss_desc_from_external_signer(network, account_type, account_num, master_blinding_key, dpk)
 
 # Account discovery
 GAP_LIMIT = 20